International US Dvorak with dead keys
Mapping us(dvorak-intl) now has dead keys in a way similar to us(intl).  The
symbols mapping is directly taken from the latter, moving the keys' symbols for
all modifier as a whole according to their position in the Dvorak mapping.

This allows for identical mappings for us(dvorak-intl) with a "normal keyboard"
and us(intl) used with a keyboard which remaps its keycode in hardware (e.g.
TypeMatrix 2030).

Signed-off-by: Olivier Mehani <shtrom@ssji.net>