Basic dvorak mappings from us in macintosh_vndr/us
Support for variants dvorak, dvorak-intl and dvorak-alt-intl added in
macintosh_vndr/us by just including macintosh_vndr/us(extended) and us(%v).

XXX: This should rather be done in rules/base.ml_s.part with an entry like
 $macs          *       =       pc+macintosh_vndr/us(extended)+%l%(v)
so that not only these three Dvorak mappings are supported. However, I
couldn't find a clean way to avoid the already existing rule
 $macs          $macvendorlayouts       =       pc+macintosh_vndr/%l%(v)
to apply as "us" is in $macvendorlayouts and the variant is not taken into
account.

Signed-off-by: Olivier Mehani <shtrom@ssji.net>